Move signature text grammars into CSharpText
Move canonical signatures, XML-documentation identity notation, and primitive type names into the dependency-free CSharpText leaf. Keep MemberAnchor and metadata-model adaptation in their existing ownership layers.


Copilot-Session: bc2613de-a721-4efd-8390-56db684eefd7
