+++ src/ILInspector.Decompiler/ILInspector.Decompiler.csproj
@@ -23,6 +23,7 @@
 
   <ItemGroup>
     <Compile Include="..\UnionPolyfill.cs" Link="UnionPolyfill.cs" />
+    <ProjectReference Include="..\CSharpText\CSharpText.csproj" />
     <ProjectReference Include="..\ILInspector.CSharp\ILInspector.CSharp.csproj" />
     <ProjectReference Include="..\ILInspector.Findings\ILInspector.Findings.csproj" />
     <ProjectReference Include="..\ILInspector.Metadata\ILInspector.Metadata.csproj" />
