+++ src/ILInspector.Metadata/ApiSurfaceExtractor.cs
@@ -4,6 +4,7 @@ using System.Reflection.Metadata;
 using System.Reflection.Metadata.Ecma335;
 using System.Reflection.PortableExecutable;
 using System.Text;
+using CSharpText;
 using ILInspector.CSharp;
 
 namespace ILInspector.Metadata;
