+++ tests/ILInspector.Metadata.Tests/ApiSignatureModelTests.cs
@@ -353,80 +353,6 @@ public sealed class ApiSignatureModelTests
         Assert.False(ApiMemberIdentity.TryGetXmlDocMemberIdentity(type, member, out _));
     }
 
-    [Fact]
-    public void XmlDocParameterNormalization_UsesGenericMaps()
-    {
-        Dictionary<string, int> typeParameters = new(StringComparer.Ordinal)
-        {
-            ["TType"] = 0
-        };
-        Dictionary<string, int> methodParameters = new(StringComparer.Ordinal)
-        {
-            ["TMethod"] = 0
-        };
-
-        Assert.Equal("T0", ApiMemberIdentity.NormalizeXmlDocParameterType("TType", typeParameters, methodParameters));
-        Assert.Equal("M0", ApiMemberIdentity.NormalizeXmlDocParameterType("TMethod", typeParameters, methodParameters));
-        Assert.Equal(
-            "System.Collections.Generic.Dictionary{T0,M0}",
-            ApiMemberIdentity.NormalizeXmlDocParameterType(
-                "System.Collections.Generic.Dictionary<TType, TMethod>",
-                typeParameters,
-                methodParameters));
-    }
-
-    [Fact]
-    public void XmlDocSignatureParameterNormalization_StripsParameterNamesForFallback()
-    {
-        Assert.Equal(
-            "System.Int32",
-            ApiMemberIdentity.NormalizeXmlDocSignatureParameter(
-                "int index",
-                new Dictionary<string, int>(StringComparer.Ordinal),
-                new Dictionary<string, int>(StringComparer.Ordinal)));
-        Assert.Equal(
-            "System.Collections.Generic.Dictionary{System.String,System.Int32}",
-            ApiMemberIdentity.NormalizeXmlDocSignatureParameter(
-                "System.Collections.Generic.Dictionary<string, int> values = null",
-                new Dictionary<string, int>(StringComparer.Ordinal),
-                new Dictionary<string, int>(StringComparer.Ordinal)));
-        Assert.Equal(
-            "System.Int32",
-            ApiMemberIdentity.NormalizeXmlDocSignatureParameter(
-                """[System.ComponentModel.DefaultValue("=")] int value = 0""",
-                new Dictionary<string, int>(StringComparer.Ordinal),
-                new Dictionary<string, int>(StringComparer.Ordinal)));
-        Assert.Equal(
-            "System.Int32@",
-            ApiMemberIdentity.NormalizeXmlDocSignatureParameter(
-                "[System.Runtime.InteropServices.In] ref int value",
-                new Dictionary<string, int>(StringComparer.Ordinal),
-                new Dictionary<string, int>(StringComparer.Ordinal)));
-        Assert.Equal(
-            "System.Int32",
-            ApiMemberIdentity.NormalizeXmlDocSignatureParameter(
-                """[System.ComponentModel.DefaultValue("=")] int value = 0""",
-                new Dictionary<string, int>(StringComparer.Ordinal),
-                new Dictionary<string, int>(StringComparer.Ordinal)));
-    }
-
-    [Theory]
-    [InlineData("ref int", "System.Int32@")]
-    [InlineData("out int", "System.Int32@")]
-    [InlineData("in int", "System.Int32@")]
-    [InlineData("System.Int32@", "System.Int32@")]
-    [InlineData("params int[]", "System.Int32[]")]
-    [InlineData("int*", "System.Int32*")]
-    [InlineData("int[,]", "System.Int32[,]")]
-    [InlineData("System.Int32[0:,0:]", "System.Int32[,]")]
-    [InlineData("int?", "System.Nullable{System.Int32}")]
-    [InlineData("System.DateTime?", "System.Nullable{System.DateTime}")]
-    [InlineData("string?", "System.String")]
-    public void XmlDocParameterNormalization_PreservesByRefMarkers(string input, string expected)
-    {
-        Assert.Equal(expected, ApiMemberIdentity.NormalizeXmlDocParameterType(input));
-    }
-
     static ApiType GetType(string typeName)
         => Surface.Types.First(type => type.Name == typeName);
